qcacld-3.0: Harmonize sys_build_message_header()
The parameter identifiers used in the implementation of
sys_build_message_header() differ from the ones used in the
prototype. Since this can be confusing for both humans and code
analysis tools, harmonize on a good set of identifiers. In addition,
consolidate the two distince documentation blocks so that it is the
interface that is fully and correctly documented, and not the
implementation.

Change-Id: Iaeee04ec918fad20a43c5804f9e50b7b208dfb94
CRs-Fixed: 2376109