qcacld-3.0: Remove unused logic to convert DFS channels to active
lim.dfschannelList.timeStamp is a array of
SIR_MAX_24G_5G_CHANNEL_RANGE(166). In some countries channel greater
than 165 are allowed and thus if channel is greater than 165 driver
access out of bound memory which lead to corruption of
lim.sme_msg_callback.

lim.dfschannelList.timeStamp is used to convert DFS channels to
active, which is used for host handled scan to decide if active
scan can be done on DFS channel. Scan now been handled in firmware
this logic is unused and can be removed.

Change-Id: Ib301fa70a6b7ed99f3ecacbf344dd9e821079149
CRs-Fixed: 2374977