Revert "net: qrtr: Send HELLO message on endpoint register"
This reverts commit 7abb34704b7a36ac5f01e69277e2c914eaf990d8.

Originally MHI Host API's assumed the client would not send until it
first received a packet from the device. This assumption is broken with
the change to queue hello work.

The MHI driver should split the channel initialization so that probe
occurs after the UL channel is started. This will allow the QRTR driver
to send immediately after probe. Revert this change until the channel
initialization change from MHI is ready.

Change-Id: I174d8c204eb1c36b0519005f15d79080fed8139f
Signed-off-by: Chris Lew <clew@codeaurora.org>