net: macsec: introduce the macsec_context structure
This patch introduces the macsec_context structure. It will be used
in the kernel to exchange information between the common MACsec
implementation (macsec.c) and the MACsec hardware offloading
implementations. This structure contains pointers to MACsec specific
structures which contain the actual MACsec configuration, and to the
underlying device (netdev or phydev).

Change-Id: I9bffddd4570257fa1dad18f6ae7be943ec673c1f
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Git-commit: e0d5d9ca24ea3ee9970413f8ed69827e8cf43fbd
Git-repo: https://github.com/aquantia/linux-4.14-atlantic-forwarding
Signed-off-by: Jinesh K. Jayakumar <jineshk@codeaurora.org>